# internal/router/v1/user_router.go
package v1

import (
	"net/http"
	request "simpletodo/internal/dto/request"
	"simpletodo/internal/dto/response"
	"simpletodo/internal/middleware"
	repository2 "simpletodo/internal/repository"
	"simpletodo/internal/service"
	"simpletodo/internal/util/mapper"
	"strconv"

	"github.com/go-playground/validator/v10"
	"github.com/labstack/echo/v4"
	"gorm.io/gorm"
)

type UserController struct {
	UserService *service.UserService
}

func NewUserController(s *service.UserService) *UserController {
	return &UserController{UserService: s}
}

func (uc *UserController) getUserByID(c echo.Context) error {
	idStr := c.Param("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid ID", err.Error(), true)
	}
	user, err := uc.UserService.GetByID(uint(id))
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusNotFound, "User not found", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "User fetched successfully", user, false)
}

func (uc *UserController) getAllUsers(c echo.Context) error {
	pagination, err := validatePagination(c)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Bad request error", err.Error(), true)
	}
	users, err := uc.UserService.GetAll(pagination)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Error", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "Users fetched successfully", users, false)
}

func (uc *UserController) updateUser(c echo.Context) error {
	idStr := c.Param("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid ID", err.Error(), true)
	}
	var body request.AdminUpdateUserRequest
	if err := bindAndValidate(c, &body); err != nil {
		return err
	}
	updatedUser, err := uc.UserService.UpdateByAdmin(uint(id), body)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Failed to update", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "User updated successfully", updatedUser, false)
}

func (uc *UserController) deleteUser(c echo.Context) error {
	idStr := c.Param("id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid ID", err.Error(), true)
	}
	user, err := uc.UserService.GetByID(uint(id))

	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Error fetching user", err.Error(), true)
	}

	if user == nil {
		return response.WriteJSONResponse(c, http.StatusNotFound, "User not found", nil, true)
	}

	if user.Role == "ADMIN" {
		return response.WriteJSONResponse(c, http.StatusForbidden, "Cannot delete admin user", nil, true)
	}

	err = uc.UserService.Delete(uint(id))
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Error deleting user", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "User deleted successfully", nil, false)
}

func (uc *UserController) getUserProfile(c echo.Context) error {
	id := c.Get("user_id").(float64)
	user, err := uc.UserService.GetByID(uint(id))
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusNotFound, "User not found", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "User profile fetched successfully", user, false)
}

func (uc *UserController) updateProfile(c echo.Context) error {
	id := c.Get("user_id").(float64)
	var body request.UpdateUserRequest
	if err := bindAndValidate(c, &body); err != nil {
		return err
	}
	updatedUser, err := uc.UserService.UpdateProfile(uint(id), body)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Failed to update", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "User updated successfully", updatedUser, false)
}

func (uc *UserController) getAISettings(c echo.Context) error {
	id := c.Get("user_id").(float64)
	settings, err := uc.UserService.GetAISettings(uint(id))
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Failed to fetch AI settings", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "AI Settings fetched", settings, false)
}

func (uc *UserController) updateAISettings(c echo.Context) error {
	id := c.Get("user_id").(float64)
	var body request.UpdateAISettingsRequest
	if err := c.Bind(&body); err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid data", err.Error(), true)
	}

	validate := validator.New()
	if err := validate.Struct(body); err != nil {
		var errorsString []string
		for _, e := range err.(validator.ValidationErrors) {
			errorsString = append(errorsString, e.Field()+" is "+e.Tag()+" "+e.Param())
		}
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid request", errorsString, true)
	}

	settings, err := uc.UserService.UpdateAISettings(uint(id), body)
	if err != nil {
		return response.WriteJSONResponse(c, http.StatusInternalServerError, "Failed to update AI settings", err.Error(), true)
	}
	return response.WriteJSONResponse(c, http.StatusOK, "AI Settings updated", settings, false)
}

func bindAndValidate(c echo.Context, body any) error {
	if err := c.Bind(body); err != nil {
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid", err.Error(), true)
	}
	validate := validator.New()
	errorValidate := validate.Struct(body)
	if errorValidate != nil {
		var errorsString []string
		for _, e := range errorValidate.(validator.ValidationErrors) {
			errorsString = append(errorsString, e.Field()+" is "+e.Tag()+" "+e.Param())
		}
		return response.WriteJSONResponse(c, http.StatusBadRequest, "Invalid request", errorsString, true)
	}

	return nil
}

func UserRouters(db *gorm.DB, v1 *echo.Group) {
	userRepository := repository2.NewUserRepository(db)
	aiServerRepository := repository2.NewAIServerRepository(db)
	userMapper := mapper.NewUserMapperImpl()

	userService := service.NewUserService(userRepository, aiServerRepository, userMapper)
	userController := NewUserController(userService)

	usersGroup := v1.Group("/users")
	usersGroup.Use(middleware.JWTMiddleware)
	usersGroup.Use(middleware.AdminOnlyMIddleware)

	userProfileGroup := v1.Group("/profile")
	userProfileGroup.Use(middleware.JWTMiddleware)

	// @Summary Get all users
	// @Description Retrieve a paginated list of all users (Admin only)
	// @Tags Users
	// @Security BearerAuth
	// @Param page query int false "Page number"
	// @Param limit query int false "Items per page"
	// @Param sort query string false "Sort order" Enums(asc, desc) default(asc)
	// @Param search query string false "Search by name, email or username"
	// @Success 200 {object} response.UsersResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 403 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /users [get]
	usersGroup.GET("", userController.getAllUsers)

	// @Summary Get user by ID
	// @Description Retrieve details of a specific user by ID (Admin only)
	// @Tags Users
	// @Security BearerAuth
	// @Param id path int true "User ID"
	// @Success 200 {object} response.UserResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 403 {object} response.StandardResponseError
	// @Failure 404 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /users/user/{id} [get]
	usersGroup.GET("/user/:id", userController.getUserByID)

	// @Summary Delete user
	// @Description Delete a specific user by ID (Admin only, cannot delete admin users)
	// @Tags Users
	// @Security BearerAuth
	// @Param id path int true "User ID"
	// @Success 200 {object} response.EmptyResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 403 {object} response.StandardResponseError
	// @Failure 404 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /users/user/{id} [delete]
	usersGroup.DELETE("/user/:id", userController.deleteUser)

	// @Summary Update user
	// @Description Update a user's name, avatar and optionally reset their password. The email address cannot be changed (Admin only).
	// @Tags Users
	// @Security BearerAuth
	// @Param id path int true "User ID"
	// @Param user body request.AdminUpdateUserRequest true "Administrative user update payload (email is immutable)"
	// @Success 200 {object} response.UserResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 403 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /users/user/{id} [put]
	usersGroup.PUT("/user/:id", userController.updateUser)

	// @Summary Get current user profile
	// @Description Retrieve the profile of the authenticated user
	// @Tags Profile
	// @Security BearerAuth
	// @Success 200 {object} response.UserResponse
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 404 {object} response.StandardResponseError
	// @Router /profile [get]
	userProfileGroup.GET("", userController.getUserProfile)

	// @Summary Update current user profile
	// @Description Update the profile of the authenticated user
	// @Tags Profile
	// @Security BearerAuth
	// @Accept json
	// @Produce json
	// @Param user body request.UpdateUserRequest true "Profile update payload"
	// @Success 200 {object} response.UserResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /profile [patch]
	userProfileGroup.PATCH("", userController.updateProfile)

	// @Summary Get AI Settings
	// @Description Get the current user's AI server configuration
	// @Tags Profile
	// @Security BearerAuth
	// @Success 200 {object} response.AISettingsResponse
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /profile/ai-settings [get]
	userProfileGroup.GET("/ai-settings", userController.getAISettings)

	// @Summary Update AI Settings
	// @Description Update the current user's AI server configuration
	// @Tags Profile
	// @Security BearerAuth
	// @Param payload body request.UpdateAISettingsRequest true "AI Settings payload"
	// @Success 200 {object} response.AISettingsResponse
	// @Failure 400 {object} response.StandardResponseError
	// @Failure 401 {object} response.StandardResponseError
	// @Failure 500 {object} response.StandardResponseError
	// @Router /profile/ai-settings [put]
	userProfileGroup.PUT("/ai-settings", userController.updateAISettings)
}